+++ optd-persistent/src/migrator/cost_model/m20241029_000001_attribute_stats_junction.rs
@@ -1,13 +1,20 @@
-use super::constraint::Constraint;
-use super::table_attribute::TableAttribute;
+/*
+Table attribute_stats_junction {
+  attr_id integer [ref: > table_attribute.id]
+  stats_id integer [ref: > attribute_stats.id]
+}
+ */
+
+use crate::migrator::cost_model::attribute_stat::AttributeStat;
+use crate::migrator::cost_model::table_attribute::TableAttribute;
 use sea_orm_migration::prelude::*;
-use sea_orm_migration::schema::integer;
+use sea_orm_migration::schema::*;
 
 #[derive(Iden)]
-pub enum ForeignConstraintRefAttributeJunction {
+pub enum AttributeStatsJunction {
     Table,
-    ConstraintId,
     AttrId,
+    StatsId,
 }
 
 #[derive(DeriveMigrationName)]
@@ -19,44 +26,47 @@ impl MigrationTrait for Migration {
         manager
             .create_table(
                 Table::create()
-                    .table(ForeignConstraintRefAttributeJunction::Table)
+                    .table(AttributeStatsJunction::Table)
                     .if_not_exists()
-                    .col(integer(ForeignConstraintRefAttributeJunction::ConstraintId))
-                    .col(integer(ForeignConstraintRefAttributeJunction::AttrId))
+                    .col(integer(AttributeStatsJunction::AttrId).not_null())
+                    .col(integer(AttributeStatsJunction::StatsId).not_null())
                     .foreign_key(
                         ForeignKey::create()
                             .from(
-                                ForeignConstraintRefAttributeJunction::Table,
-                                ForeignConstraintRefAttributeJunction::ConstraintId,
-                            )
-                            .to(Constraint::Table, Constraint::Id)
-                            .on_delete(ForeignKeyAction::Cascade)
-                            .on_update(ForeignKeyAction::Cascade),
-                    )
-                    .foreign_key(
-                        ForeignKey::create()
-                            .from(
-                                ForeignConstraintRefAttributeJunction::Table,
-                                ForeignConstraintRefAttributeJunction::AttrId,
+                                AttributeStatsJunction::Table,
+                                AttributeStatsJunction::AttrId,
                             )
                             .to(TableAttribute::Table, TableAttribute::Id)
                             .on_delete(ForeignKeyAction::Cascade)
                             .on_update(ForeignKeyAction::Cascade),
                     )
+                    .foreign_key(
+                        ForeignKey::create()
+                            .from(
+                                AttributeStatsJunction::Table,
+                                AttributeStatsJunction::StatsId,
+                            )
+                            .to(AttributeStat::Table, AttributeStat::Id)
+                            .on_delete(ForeignKeyAction::Cascade)
+                            .on_update(ForeignKeyAction::Cascade),
+                    )
                     .primary_key(
                         Index::create()
-                            .col(ForeignConstraintRefAttributeJunction::ConstraintId)
-                            .col(ForeignConstraintRefAttributeJunction::AttrId),
+                            .col(AttributeStatsJunction::AttrId)
+                            .col(AttributeStatsJunction::StatsId)
+                            .name("attribute_stats_junction_pk")
+                            .unique(),
                     )
                     .to_owned(),
             )
             .await
     }
+
     async fn down(&self, manager: &SchemaManager) -> Result<(), DbErr> {
         manager
             .drop_table(
                 Table::drop()
-                    .table(ForeignConstraintRefAttributeJunction::Table)
+                    .table(AttributeStatsJunction::Table)
                     .to_owned(),
             )
             .await
