+++ optd-persistent/src/migrator/memo/m20241029_000001_logical_expression.rs
@@ -1,43 +1,44 @@
+use crate::migrator::memo::cascades_group::CascadesGroup;
 use sea_orm_migration::{prelude::*, schema::*};
 
-use super::physical_expression::PhysicalExpression;
-
 #[derive(DeriveIden)]
-pub enum CascadesGroup {
+pub enum LogicalExpression {
     Table,
     Id,
-    LatestWinner,
-    InProgress,
-    IsOptimized,
+    GroupId,
+    Fingerprint,
+    Data,
 }
 
 pub struct Migration;
 
 impl MigrationName for Migration {
     fn name(&self) -> &str {
-        "m20241029_000001_cascades_group"
+        "m20241029_000001_logical_expression"
     }
 }
 
 #[async_trait::async_trait]
 impl MigrationTrait for Migration {
+    // TODO add expression / root operator variant identifier
     async fn up(&self, manager: &SchemaManager) -> Result<(), DbErr> {
         manager
             .create_table(
                 Table::create()
-                    .table(CascadesGroup::Table)
+                    .table(LogicalExpression::Table)
                     .if_not_exists()
-                    .col(pk_auto(CascadesGroup::Id))
-                    .col(integer_null(CascadesGroup::LatestWinner)) // TODO foreign key
-                    .col(boolean(CascadesGroup::InProgress))
-                    .col(boolean(CascadesGroup::IsOptimized))
+                    .col(pk_auto(LogicalExpression::Id))
+                    .col(integer(LogicalExpression::GroupId))
                     .foreign_key(
                         ForeignKey::create()
-                            .from(CascadesGroup::Table, CascadesGroup::LatestWinner)
-                            .to(PhysicalExpression::Table, PhysicalExpression::Id)
+                            .name("fk-logical_expression-group_id")
+                            .from(LogicalExpression::Table, LogicalExpression::GroupId)
+                            .to(CascadesGroup::Table, CascadesGroup::Id)
                             .on_delete(ForeignKeyAction::Cascade)
                             .on_update(ForeignKeyAction::Cascade),
                     )
+                    .col(big_unsigned(LogicalExpression::Fingerprint))
+                    .col(json(LogicalExpression::Data))
                     .to_owned(),
             )
             .await
@@ -45,7 +46,7 @@ impl MigrationTrait for Migration {
 
     async fn down(&self, manager: &SchemaManager) -> Result<(), DbErr> {
         manager
-            .drop_table(Table::drop().table(CascadesGroup::Table).to_owned())
+            .drop_table(Table::drop().table(LogicalExpression::Table).to_owned())
             .await
     }
 }
