+++ optd-persistent/src/migrator/memo/m20241029_000001_logical_group_junction.rs
@@ -1,4 +1,4 @@
-use super::{cascades_group::CascadesGroup, logical_expression::LogicalExpression};
+use crate::migrator::memo::{cascades_group::CascadesGroup, logical_expression::LogicalExpression};
 use sea_orm_migration::{prelude::*, schema::*};
 
 #[derive(DeriveIden)]
