+++ optd-persistent/src/migrator/memo/m20241029_000001_physical_expression.rs
@@ -1,4 +1,4 @@
-use super::cascades_group::CascadesGroup;
+use crate::migrator::memo::cascades_group::CascadesGroup;
 use sea_orm_migration::{prelude::*, schema::*};
 
 #[derive(DeriveIden)]
