+++ optd-persistent/src/migrator/memo/m20241029_000001_physical_property.rs
@@ -1,4 +1,4 @@
-use super::physical_expression::PhysicalExpression;
+use crate::migrator::memo::physical_expression::PhysicalExpression;
 use sea_orm_migration::{prelude::*, schema::*};
 
 #[derive(DeriveIden)]
