+++ optd-persistent/src/migrator/mod.rs
@@ -1,48 +1,7 @@
 use sea_orm_migration::prelude::*;
 
-mod m20241029_000001_attribute_stat;
-mod m20241029_000001_attribute_stats_junction;
-mod m20241029_000001_cascades_group;
-mod m20241029_000001_constraint;
-mod m20241029_000001_constraint_attribute_junction;
-mod m20241029_000001_cost;
-mod m20241029_000001_database_metadata;
-mod m20241029_000001_event;
-mod m20241029_000001_foreign_constraint_ref_attribute_junction;
-mod m20241029_000001_group_winner;
-mod m20241029_000001_index;
-mod m20241029_000001_logical_expression;
-mod m20241029_000001_logical_group_junction;
-mod m20241029_000001_logical_property;
-mod m20241029_000001_namespace_metadata;
-mod m20241029_000001_physical_expression;
-mod m20241029_000001_physical_group_junction;
-mod m20241029_000001_physical_property;
-mod m20241029_000001_table_attribute;
-mod m20241029_000001_table_metadata;
-mod m20241029_000001_trigger;
-
-use m20241029_000001_attribute_stat as attribute_stat;
-use m20241029_000001_attribute_stats_junction as attribute_stats_junction;
-use m20241029_000001_cascades_group as cascades_group;
-use m20241029_000001_constraint as constraint;
-use m20241029_000001_constraint_attribute_junction as constraint_attribute_junction;
-use m20241029_000001_cost as cost;
-use m20241029_000001_database_metadata as database_metadata;
-use m20241029_000001_event as event;
-use m20241029_000001_foreign_constraint_ref_attribute_junction as foreign_constraint_ref_attribute_junction;
-use m20241029_000001_group_winner as group_winner;
-use m20241029_000001_index as index;
-use m20241029_000001_logical_expression as logical_expression;
-use m20241029_000001_logical_group_junction as logical_group_junction;
-use m20241029_000001_logical_property as logical_property;
-use m20241029_000001_namespace_metadata as namespace_metadata;
-use m20241029_000001_physical_expression as physical_expression;
-use m20241029_000001_physical_group_junction as physical_group_junction;
-use m20241029_000001_physical_property as physical_property;
-use m20241029_000001_table_attribute as table_attribute;
-use m20241029_000001_table_metadata as table_metadata;
-use m20241029_000001_trigger as trigger;
+mod cost_model;
+mod memo;
 
 pub struct Migrator;
 
@@ -50,28 +9,27 @@ pub struct Migrator;
 impl MigratorTrait for Migrator {
     fn migrations() -> Vec<Box<dyn MigrationTrait>> {
         vec![
-            Box::new(cascades_group::Migration),
-            Box::new(logical_expression::Migration),
-            Box::new(logical_group_junction::Migration),
-            Box::new(logical_property::Migration),
-            Box::new(physical_expression::Migration),
-            Box::new(physical_group_junction::Migration),
-            Box::new(physical_property::Migration),
-
-            Box::new(table_metadata::Migration),
-            Box::new(table_attribute::Migration),
-            Box::new(index::Migration),
-            Box::new(namespace_metadata::Migration),
-            Box::new(database_metadata::Migration),
-            Box::new(event::Migration),
-            Box::new(cost::Migration),
-            Box::new(group_winner::Migration),
-            Box::new(attribute_stat::Migration),
-            Box::new(attribute_stats_junction::Migration),
-            Box::new(constraint::Migration),
-            Box::new(constraint_attribute_junction::Migration),
-            Box::new(trigger::Migration),
-            Box::new(foreign_constraint_ref_attribute_junction::Migration),
+            Box::new(memo::cascades_group::Migration),
+            Box::new(memo::group_winner::Migration),
+            Box::new(memo::logical_expression::Migration),
+            Box::new(memo::logical_group_junction::Migration),
+            Box::new(memo::logical_property::Migration),
+            Box::new(memo::physical_expression::Migration),
+            Box::new(memo::physical_group_junction::Migration),
+            Box::new(memo::physical_property::Migration),
+            Box::new(cost_model::database_metadata::Migration),
+            Box::new(cost_model::namespace_metadata::Migration),
+            Box::new(cost_model::table_metadata::Migration),
+            Box::new(cost_model::table_attribute::Migration),
+            Box::new(cost_model::attribute_stat::Migration),
+            Box::new(cost_model::constraint::Migration),
+            Box::new(cost_model::attribute_stats_junction::Migration),
+            Box::new(cost_model::constraint_attribute_junction::Migration),
+            Box::new(cost_model::foreign_constraint_ref_attribute_junction::Migration),
+            Box::new(cost_model::index::Migration),
+            Box::new(cost_model::event::Migration),
+            Box::new(cost_model::cost::Migration),
+            Box::new(cost_model::trigger::Migration),
         ]
     }
 }
